+++ src/app.module.ts
@@ -2,20 +2,20 @@ import { Module } from '@nestjs/common';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { UserModule } from './user/user.module';
-import { ArtistModule } from './artist/artist.module';
-import { TrackModule } from './track/track.module';
-import { AlbumModule } from './album/album.module';
-import { FavoritesModule } from './favorites/favorites.module';
-import { TrackService } from './track/track.service';
-import { AlbumService } from './album/album.service';
-import { TrackController } from './track/track.controller';
-import { AlbumController } from './album/album.controller';
-import { AlbumEntity } from './album';
-import { ArtistEntity } from './artist';
-import { FavoritesEntity } from './favorites';
-import { TrackEntity } from './track';
-import { UserEntity } from './user';
+import { UserModule } from './modules/user/user.module';
+import { ArtistModule } from './modules/artist/artist.module';
+import { TrackModule } from './modules/track/track.module';
+import { AlbumModule } from './modules/album/album.module';
+import { FavoritesModule } from './modules/favorites/favorites.module';
+import { TrackService } from './modules/track/track.service';
+import { AlbumService } from './modules/album/album.service';
+import { TrackController } from './modules/track/track.controller';
+import { AlbumController } from './modules/album/album.controller';
+import { AlbumEntity } from './modules/album';
+import { ArtistEntity } from './modules/artist';
+import { FavoritesEntity } from './modules/favorites';
+import { TrackEntity } from './modules/track';
+import { UserEntity } from './modules/user';
 @Module({
   imports: [
     TypeOrmModule.forRoot({
