+++ src/modules/favorites/index.ts
