+++ src/modules/track/track.entity.ts
@@ -1,5 +1,5 @@
-import { AlbumEntity } from 'src/album/album.entity';
-import { ArtistEntity } from 'src/artist/artist.entity';
+import { AlbumEntity } from 'src/modules/album/album.entity';
+import { ArtistEntity } from 'src/modules/artist/artist.entity';
 import {
   Entity,
   PrimaryGeneratedColumn,
